Make formatter an interface, with DB implementation
Also add a test fake implementation which we can use in unit testing.
